refactor: (#70) ApiResponseDto 파일 통합 및 공통 경로 적용
